Extract encoded message with secondary locations to utility file
